site: Improve mobile story and better links
