+++ package.json
@@ -5,7 +5,6 @@
   "main": "gulpfile.babel.js",
   "dependencies": {},
   "devDependencies": {
-    "Bootstrap-Offcanvas": "github:iamphill/Bootstrap-Offcanvas",
     "async": "^1.5.0",
     "babel-core": "^6.1.2",
     "babel-eslint": "^5.0.0-beta6",
@@ -34,7 +33,7 @@
     "pre-commit": "^1.1.2",
     "require-dir": "^0.3.0",
     "run-sequence": "^1.1.5",
-    "stickyfill": "^1.1.1",
+    "slicknav": "^1.0.6",
     "tether": "^1.1.1",
     "webpack": "^1.12.2",
     "webpack-stream": "^3.1.0"
