+++ site/public/_header.jade
@@ -4,10 +4,4 @@ nav.d-header.navbar.navbar-fixed-top.navbar-dark.bg-inverse
       a.navbar-brand
         | IPFS
         span.d-header-light  distributions
-
-      button#d-toggle.navbar-toggler.offcanvas-toggle.hidden-sm-up.pull-right(
-        type='button'
-        data-toggle='offcanvas'
-        data-target='#d-navbar-offcanvas'
-      )
-        | &#9776;
+      #d-header-menu
