+++ site/public/_js/script.js
@@ -1,49 +1,16 @@
 const $ = window.jQuery = require('jquery')
 window.Tether = require('tether')
-const Stickyfill = require('stickyfill')()
 require('bootstrap')
-require('bootstrap-offcanvas/dist/js/bootstrap.offcanvas.js')
-const platform = require('platform')
+require('slicknav/jquery.slicknav.js')
 
-Stickyfill.add($('.sticky')[0])
+const handlePlatform = require('./_platform')
 
-function getArch (os, arch) {
-  if (arch === '32' && os !== 'darwin') return '386'
-  return 'amd64'
-}
+$(() => {
+  handlePlatform()
 
-function getOsName (family) {
-  family = family.toLowerCase()
-
-  if (family.match(/windows/)) return 'windows'
-  if (family.match(/(os x|ios)/)) return 'darwin'
-  return 'linux'
-}
-
-function buildDownloadLink (id, version, os) {
-  const osName = getOsName(os.family)
-  const arch = getArch(osName, os.architecture)
-
-  return `${id}/${version}/${osName}-${arch}/${id}.zip`
-}
-
-$('.d-component-download-btn').each(function () {
-  const elem = $(this)
-  const version = elem.data('version')
-  const id = elem.data('id')
-
-  const link = buildDownloadLink(id, version, platform.os)
-  let osName = getOsName(platform.os.family)
-  let arch = getArch(osName, platform.os.architecture)
-
-  if (arch === 'amd64') {
-    arch = '64bit'
-  } else {
-    arch = '32bit'
-  }
-
-  if (osName === 'darwin') osName = 'OS X'
-
-  elem.attr('href', link)
-  elem.parent().find('.d-component-arch').text(`Version ${version} for ${osName} ${arch}`)
+  $('#d-navbar').slicknav({
+    label: '',
+    closeOnClick: true,
+    prependTo: '#d-header-menu'
+  })
 })
