+++ site/public/css/style.less
@@ -1,6 +1,7 @@
 @import (inline) '../../../node_modules/bootstrap/dist/css/bootstrap.css';
-@import (inline) '../../../node_modules/bootstrap-offcanvas/dist/css/bootstrap.offcanvas.css';
+@import (inline) '../../../node_modules/slicknav/dist/slicknav.css';
 
+@import './_mixins';
 @import './_variables';
 @import './_base';
